+++ CHANGELOG.md
@@ -7,6 +7,10 @@ CHANGELOG
     - cf. `preview(...)` is a one-off action that doesn't change the default
       preview command
 - Added `change-preview-window(...)` action
+    - You can rotate through the different options separated by `|`
+      ```sh
+      fzf --preview 'cat {}' --bind 'ctrl-/:change-preview-window(right,70%|down,40%,border-horizontal|hidden|right)'
+      ```
 
 0.28.0
 ------